Removes all uses of unified memory in CUDA-MPI
These were causing a lot of instability in the tests. Plus, they were
killing the performance of single GPU / multiple process tests.

Specifically:

+ `device::Var`` is now implemented with device allocations and explicit
  copies.
+ Moves all message buffers to device (not unified) memory.
+ This breaks support for "not CUDA aware" MPI.

Also:

+ Implements `host::memcpy_todevice()``
+ Moves Particles `opt_*` parameters from `#define` to `constexpr`
+ Updates `benchmark` script
+ `Particles::update_tile_info()` was optimized with shared memory

+++ cuda-mpi/species.h
@@ -112,7 +112,8 @@ private:
     double * d_energy;
 
     /// @brief Total number of particles moved
-    uint64_t * d_nmove;
+    /// @note CUDA gets confused with uint64_t, don't use it
+    unsigned long long * d_nmove;
 
     /**
      * @brief Shift particle positions due to moving window motion
@@ -154,7 +155,7 @@ private:
 
 public:
 
-     /// @brief Species name
+    /// @brief Species name
     const std::string name;
 
      /// @brief  Mass over charge ratio
@@ -410,8 +411,8 @@ public:
      * 
      * @return uint64_t
      */
-    uint64_t get_nmove() const {
-        uint64_t nmove;
+    unsigned long long get_nmove() const {
+        unsigned long long nmove;
         device::memcpy_tohost( &nmove, d_nmove, 1 );
 
         return nmove;
